Improves code organization and structure
Refactors the codebase to adhere to a consistent organizational pattern across key modules, enhancing readability and maintainability.

This includes standardizing section sequences, comment separators, documentation, naming conventions, and testing patterns.

Updates the triangulation data structure for improved clarity and efficiency, including the addition of helper functions and constants, along with corresponding documentation.

Renames the optimization recommendations file to improve discoverability.
