+++ cspell.json
@@ -40,6 +40,7 @@
         "findutils",
         "finitecheck",
         "fuga",
+        "functiongenerics",
         "Getchell",
         "getrandom",
         "golangci",
