+++ src/geometry/matrix.rs
@@ -1,8 +1,16 @@
 //! Matrix operations.
 
+// =============================================================================
+// IMPORTS
+// =============================================================================
+
 use anyhow::Result;
 use peroxide::prelude::*;
 use thiserror::Error;
+
+// =============================================================================
+// ERROR TYPES
+// =============================================================================
 
 /// Error type for matrix operations.
 #[derive(Debug, Error)]
@@ -11,6 +19,10 @@ pub enum MatrixError {
     #[error("Matrix is singular!")]
     SingularMatrix,
 }
+
+// =============================================================================
+// MATRIX OPERATIONS
+// =============================================================================
 
 /// Inverts a matrix.
 ///
@@ -46,9 +58,12 @@ pub fn invert(matrix: &Matrix) -> Result<Matrix, anyhow::Error> {
     Ok(inv)
 }
 
+// =============================================================================
+// TESTS
+// =============================================================================
+
 #[cfg(test)]
 mod tests {
-
     use peroxide::c;
     use peroxide::fuga::*;
