+++ src/geometry/point.rs
@@ -266,7 +266,7 @@ mod tests {
     }
 
     // =============================================================================
-    // BASIC POINT FUNCTIONALITY
+    // BASIC POINT CREATION TESTS
     // =============================================================================
 
     #[test]
@@ -417,6 +417,10 @@ mod tests {
             epsilon = 1e-9
         );
     }
+
+    // =============================================================================
+    // HASH AND EQUALITY TESTS
+    // =============================================================================
 
     #[test]
     fn point_hash() {
@@ -723,6 +727,10 @@ mod tests {
         // Verify the original point is still available after reference conversion
         assert_relative_eq!(point_ref.to_array().as_slice(), [4.0, 5.0, 6.0].as_slice());
     }
+
+    // =============================================================================
+    // VALIDATION TESTS
+    // =============================================================================
 
     #[test]
     fn point_is_valid_f64() {
